+++ changelog.md
@@ -17,7 +17,8 @@
 + The 'step' class was renamed to 'stepy-step';
 + The '{id}-titles-{index}' was renamed to '{id}-head-{index}';
 + All callbacks now receives the variable `this` as DOM element instead jQuery;
-+ The 'current-step' class was renamed to 'stepy-active'.
++ The 'current-step' class was renamed to 'stepy-active';
++ The finish button now must be the class 'stepy-finish' instead just 'finish'.
 
 ## Fixes